refactor: split test_phases.py into phase-specific modules
